% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{createHelixAlignments}
\alias{createHelixAlignments}
\title{Sequence Alignment of Template Rhodopsin to the Query Sequences}
\usage{
createHelixAlignments(template, sequences, rcsb_id)
}
\arguments{
\item{template}{AAStringSet of template rhodopsin to get helices from}

\item{sequences}{AAStringSet of query sequences to align to the template
rhodopsin's helices}

\item{rcsb_id}{the RCSB accession code for the template rhodopsin to be used}
}
\value{
a list of AAStringSet objects containing the combined pairwise alignments
of each template's helix. Each index in the list corresponds to the helix where
that alignment was made e.g. the AAStringSet at index 1 corresponds to the combined
pairwise alignments of all the query sequences against Helix 1
}
\description{
A function that performs pairwise alignment between each of the template
rhodopsin's helices and each of the query sequences, allowing users to see
the conservation of residues across a given helix.
}
\details{
This function mimics the template-based alignment performed in the del Val
et al. (2014) study using T-Coffee which is an alignment algorithm that allows
PDB structures to be used as a 'template' when aligning. Since this is not
available in R and there are no similar R packages that can implement it, this
function performs a similar approach by identifying the helices in a given
structure and performing pairwise alignments to each one with the sequence queries.
}
\examples{
template <- template_rhodopsins[1]
sequences <- sample_rhodopsins
rcsb_id <- "1QHJ"
results <- createHelixAlignments(template = template, sequences = sequences, rcsb_id = rcsb_id)

}
\references{
Aboyoun P, Gentleman R (2024). \emph{pwalign: Perform pairwise sequence alignments}.
doi:10.18129/B9.bioc.pwalign \url{https://doi.org/10.18129/B9.bioc.pwalign}, R
package version 1.2.0, \url{https://bioconductor.org/packages/pwalign}.

del Val, C., Royuela-Flor, J., Milenkovic, S., & Bondar, A.-N. (2014).
\emph{Channelrhodopsins: A bioinformatics perspective}. Biochimica et Biophysica Acta
(BBA) - Bioenergetics, 1837(5), 643–655. https://doi.org/10.1016/j.bbabio.2013.11.005

Pagès H, Aboyoun P, Gentleman R, DebRoy S (2024). _Biostrings: Efficient _
\emph{manipulationof biological strings}. doi:10.18129/B9.bioc.Biostrings
\url{https://doi.org/10.18129/B9.bioc.Biostrings}, R package version 2.73.1,
\url{https://bioconductor.org/packages/Biostrings}.

Wright ES (2016). _Using DECIPHER v2.0 to Analyze Big Biological _
\emph{Data in R.} The R Journal, \emph{8}(1), 352-359.
}
