% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{createPWAlignments}
\alias{createPWAlignments}
\title{PW Align Helices with Query Sequences}
\usage{
createPWAlignments(template, sequences)
}
\arguments{
\item{template}{AAStringSet of template rhodopsin to get helices from}

\item{sequences}{AAStringSet of query sequences to align to helices}
}
\description{
PW Align Helices with Query Sequences
}
\references{
Pagès H, Aboyoun P, Gentleman R, DebRoy S (2024). \emph{Biostrings: Efficient manipulation
of biological strings}. doi:10.18129/B9.bioc.Biostrings
\url{https://doi.org/10.18129/B9.bioc.Biostrings}, R package version 2.73.1,
\url{https://bioconductor.org/packages/Biostrings}.

Aboyoun P, Gentleman R (2024). \emph{pwalign: Perform pairwise sequence alignments}.doi:10.18129/B9.bioc.pwalign \url{https://doi.org/10.18129/B9.bioc.pwalign}, R package version 1.2.0, \url{https://bioconductor.org/packages/pwalign}.
}
