% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{find3dStructure}
\alias{find3dStructure}
\title{Find 3D Structure Given Amino Acid Sequence}
\usage{
find3dStructure(aaSeq)
}
\arguments{
\item{aaSeq}{An AAString object of the rhodopsin sequence that we want to find the 3D structure of}
}
\value{
dataframe with information on structures
}
\description{
Find 3D Structure Given Amino Acid Sequence
}
\examples{
fastaPath <- system.file("extdata", "rcsb_pdb_8GI8.fasta", package = "RhodopXin")
seq <- loadSequence(fastaPath)
struct_df <- find3dStructure(seq)

}
\references{
Korkmaz S, Yamasan B (2024). \emph{rPDBapi: A Comprehensive Interface for Accessing the Protein
Data Bank}. R package version 2.1.1, \url{https://CRAN.R-project.org/package=rPDBapi}.
}
