% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{findHelices}
\alias{findHelices}
\title{Find Helices Positions Given RCSB ID}
\usage{
findHelices(rcsb_id)
}
\arguments{
\item{rcsb_id}{}
}
\value{
dataframe
}
\description{
Find Helices Positions Given RCSB ID
}
\references{
Grant B.J., Rodrigues A.P.C., ElSawy K.M., McCammon J.A., Caves L.S.D. (2006). \emph{Bio3D: An R package for the comparative analysis of protein structures}. \url{http://thegrantlab.org/bio3d/}
}
