% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{findMotifs}
\alias{findMotifs}
\title{Find Helices Positions Given RCSB ID}
\usage{
findMotifs()
}
\arguments{
\item{rcsb_id}{}
}
\value{
dataframe
}
\description{
Adapted directly from rPDBapi's documentation
}
