% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{helixSegments}
\alias{helixSegments}
\title{Get Helix Segments of Sequence}
\usage{
helixSegments(template)
}
\arguments{
\item{template}{}
}
\value{
dataframe

#' @references
Pagès H, Aboyoun P, Gentleman R, DebRoy S (2024). \emph{Biostrings: Efficient manipulation
of biological strings}. doi:10.18129/B9.bioc.Biostrings
\url{https://doi.org/10.18129/B9.bioc.Biostrings}, R package version 2.73.1,
\url{https://bioconductor.org/packages/Biostrings}.
}
\description{
Get Helix Segments of Sequence
}
