% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{helixSequences}
\alias{helixSequences}
\title{Get the Sequence for Each Helix of the Template}
\usage{
helixSequences(template, rcsb_id)
}
\arguments{
\item{template}{AAStringSet of template rhodopsin to get helices from}

\item{rcsb_id}{the RCSB accession code for the template rhodopsin to be used}
}
\value{
a list of AAStringSet objects containing the sequences of the template
rhodopsin's helices.
}
\description{
A function that returns a list of AAString objects containing the sequence
info of the given template rhodopsin's helices. The order in the list
corresponds to the order the helices were found in the original sequence.
}
\references{
Pagès H, Aboyoun P, Gentleman R, DebRoy S (2024). \emph{Biostrings: Efficient manipulation
of biological strings}. doi:10.18129/B9.bioc.Biostrings
\url{https://doi.org/10.18129/B9.bioc.Biostrings}, R package version 2.73.1,
\url{https://bioconductor.org/packages/Biostrings}.
}
