% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadInput.R
\name{loadFromRCSB}
\alias{loadFromRCSB}
\title{Load Sequence Given an RCSB PDB ID}
\usage{
loadFromRCSB(rcsb_id)
}
\arguments{
\item{rcsb_id}{the RCSB accession code for the template rhodopsin to get the
sequence of}
}
\value{
AAStringSet
}
\description{
A function that takes in a RCSB pdb id and returns an AAStringSet object of
the structure's sequence
}
\references{
Pagès H, Aboyoun P, Gentleman R, DebRoy S (2024). \emph{Biostrings: Efficient manipulation
of biological strings}. doi:10.18129/B9.bioc.Biostrings
\url{https://doi.org/10.18129/B9.bioc.Biostrings}, R package version 2.73.1,
\url{https://bioconductor.org/packages/Biostrings}.
}
