% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadInput.R
\name{loadSequence}
\alias{loadSequence}
\title{Load Amino Acid Sequence Input}
\usage{
loadSequence(inputStr)
}
\arguments{
\item{inputStr}{A string input corresponding to the amino acid sequence or a
path to a FASTA file containing the amino acid sequence.}
}
\value{
AAString
}
\description{
A function that takes in a string of an amino acid corresponding to a rhodopsin
or a FASTA file path containing the sequence information and creates an AAString
object out of it
}
\examples{
# Example 1:
# Using an amino acid string as input
sampleAA <- "MPFYDSRPPEGWPKGSINDMDYPLLGSICAVCCVFVAGSGIWMLYRLDLGMGYSCKPYKSGRA"
seq <- loadSequence(sampleAA)

# Example 2:
# Using a fasta file as input

# Define the path to the fasta file (this one is included in the package)
fastaPath <- system.file("extdata", "rcsb_pdb_8GI8.fasta", package = "RhodopXin")
seq <- loadSequence(fastaPath)

}
