% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{loadStructure}
\alias{loadStructure}
\title{Display 3D structure Given RCSB ID}
\usage{
loadStructure(rcsb_id)
}
\arguments{
\item{rcsb_id}{desc}
}
\value{
Viewer
}
\description{
Adapted directly from rPDBapi's documentation
}
\references{
Korkmaz S, Yamasan B (2024). \emph{rPDBapi: A Comprehensive Interface for Accessing the Protein Data Bank}. R package version 2.1.1, \url{https://CRAN.R-project.org/package=rPDBapi}.

Su W, Johnston B (2021). \emph{r3dmol: Create Interactive 3D Visualizations of Molecular Data}. Rpackage version 0.1.2, \url{https://CRAN.R-project.org/package=r3dmol}.

Wickham, H., François, R., Henry, L., Müller, K., Vaughan, D. (2023). \emph{dplyr: A Grammar of Data Manipulation}. \url{https://dplyr.tidyverse.org}.
}
