% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{resolvedMapDf}
\alias{resolvedMapDf}
\title{Generate Mapping Dataframe of Resolved Structure}
\usage{
resolvedMapDf(template, rcsb_id)
}
\arguments{
\item{template}{}

\item{rcsb_id}{}
}
\description{
Generate Mapping Dataframe of Resolved Structure
}
