% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRhodopXin.R
\name{runRhodopXin}
\alias{runRhodopXin}
\title{Run the RhodopXin Shiny App}
\usage{
runRhodopXin()
}
\description{
This function runs the Shiny web application for this package,
providing an interactive interface for a common workflow provided by
RhodopXIn. This function does not return, but a webpage will be opened.
}
\examples{
\dontrun{
runRhodopXin()
}

}
