% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_rhodopsins}
\alias{sample_rhodopsins}
\title{Selected Examples of Rhodopsin Variants}
\format{
An AAStringSet object of length 3 with the sequences of the rhodopsins.
\describe{
\item{width}{Length of the rhodopsin's sequence}
\item{seq}{Amino acid sequence of the rhodopsin}
\item{names}{the FASTA header containing the UniProt accession as well as
the entry name of the rhodopsin}
}
}
\source{
del Val, C., Royuela-Flor, J., Milenkovic, S., & Bondar, A.-N. (2014).
Channelrhodopsins: A bioinformatics perspective. Biochimica et Biophysica Acta
(BBA) - Bioenergetics, 1837(5), 643–655. https://doi.org/10.1016/j.bbabio.2013.11.005

UniProt Consortium. (2023). Acop1—Archaeal-type opsin 1—Chlamydomonas reinhardtii
(Chlamydomonas smithii) | UniProtKB. Retrieved November 24, 2024, from
https://www.uniprot.org/uniprotkb/Q93WP2/entry

UniProt Consortium. (2023). Channelopsin 1—Mesostigma viride (Green alga) | UniProtKB.
Retrieved November 24, 2024, from https://www.uniprot.org/uniprotkb/F8UVI5/entry

UniProt Consortium. (2023). Channelrhodopsin-1—Volvox carteri f. Nagariensis |
UniProtKB. Retrieved November 24, 2024, from https://www.uniprot.org/uniprotkb
/B4Y103/entry
}
\usage{
sample_rhodopsins
}
\description{
This is the sample AAStringSet output produced by running
loadSequence() with the rhodopsins.fasta file located in inst/extdata. It
contains three rhodopsins selected from the del Val et al. study (2014) which
originally looked at 11 channelrhodopsins in a bioinformatic perspective. They
included the UniProt entries for the sequences which was then used to fetch the
sequencing data to construct the rhodopsins.fasta file.

It contains the amino acid sequence of the rhodopsins with the following UniProt
entries:
\enumerate{
\item{B4Y103 - Volvox carter Channelrhodopsin-1}
\item{F8UVI5 - Mesostigma viride Channelrhodopsin-1}
\item{Q93WP2 - Chlamydomonas reinhardtii Archeal-type opsin 1}
}
}
\references{
Pagès H, Aboyoun P, Gentleman R, DebRoy S (2024). \emph{Biostrings: Efficient manipulation
of biological strings}. doi:10.18129/B9.bioc.Biostrings
\url{https://doi.org/10.18129/B9.bioc.Biostrings}, R package version 2.73.1,
\url{https://bioconductor.org/packages/Biostrings}.
}
\keyword{datasets}
