% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{templateMapping}
\alias{templateMapping}
\title{Process Template Ranges Dataframe for Mapping}
\usage{
templateMapping(template_ranges, template, rcsb_id)
}
\arguments{
\item{template_ranges}{the dataframe returned by createHelixAlignments()}

\item{rcsb_id}{desc}
}
\value{
a process dataframe with the following columns
}
\description{
Process Template Ranges Dataframe for Mapping
}
