% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{validateRcsbId}
\alias{validateRcsbId}
\title{Validate Given RCSB ID}
\usage{
validateRcsbId(rcsb_id)
}
\arguments{
\item{rcsb_id}{the RCSB accession code given by the user}
}
\value{
invisible null if the rcsb id is valid, otherwise throw an appropriate error
}
\description{
A function that validates the rcsb id input given by the user so that the PDB
structure can be fetched in the RCSB database
}
\references{
Korkmaz S, Yamasan B (2024). \emph{rPDBapi: A Comprehensive Interface for Accessing}
\emph{the Protein Data Bank}. R package version 2.1.1,
\url{https://CRAN.R-project.org/package=rPDBapi}.
}
