% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{visualizeHelixMapping}
\alias{visualizeHelixMapping}
\title{Map Helix Alignment to Structure}
\usage{
visualizeHelixMapping(template_ranges, template, rcsb_id, query_num)
}
\arguments{
\item{template_ranges}{the dataframe returned by createHelixAlignments()}

\item{rcsb_id}{the template rhodopsin to map to}

\item{query_num}{to highlight}
}
\value{
r3dmol object
}
\description{
Map Helix Alignment to Structure
}
