% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{findHelices}
\alias{findHelices}
\title{Find Helices Positions Given RCSB ID}
\usage{
findHelices(rcsb_id)
}
\arguments{
\item{rcsb_id}{the RCSB accession code to find the helices from}
}
\value{
dataframe with row names corresponding to the helices found in the
structure i.e. 'Helix 1' ... 'Helix n' and two columns: 'start' for the start
position of the helix in the sequence and 'end' for the end position
}
\description{
A function that finds all the helices in a pdb structure given by an rcsb id
and then return a dataframe for the start and end positions of each helix.
}
\examples{
result <- findHelices(rcsb_id = "3UG9")

}
\references{
Grant B.J., Rodrigues A.P.C., ElSawy K.M., McCammon J.A., Caves L.S.D. (2006).
\emph{Bio3D: An R package for the comparative analysis of protein structures}.
\url{http://thegrantlab.org/bio3d/}
}
