% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{getPDBstruct}
\alias{getPDBstruct}
\title{Retrieve PDB file from RCSB}
\usage{
getPDBstruct(rcsb_id)
}
\arguments{
\item{rcsb_id}{the RCSB accession code given by the user}
}
\value{
pdb class (see ?rPDBapi::read.pdb for specific info on components)
}
\description{
A function to retrieve the pdb file given an rcsb id. Stores the pdb file in
a temporary directory. The package deals with all this so the user does not have
to manually manage and download the pdb files themselves.
}
\references{
Grant B.J., Rodrigues A.P.C., ElSawy K.M., McCammon J.A., Caves L.S.D. (2006).
\emph{Bio3D: An R package for the comparative analysis of protein structures}.
\url{http://thegrantlab.org/bio3d/}

OpenAI. (2024). ChatGPT (November 2024 version). Retrieved from
https://chat.openai.com
}
