% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{visualizeHelixAlignments}
\alias{visualizeHelixAlignments}
\title{Visualize Helix Alignments}
\usage{
visualizeHelixAlignments(all_pwa)
}
\arguments{
\item{all_pwa}{a list of AAStringSet objects containing the combined pairwise
alignments of each template's helix. Each index in the list corresponds to the
helix where that alignment was made e.g. the AAStringSet at index 1 corresponds
to the combined pairwise alignments of all the query sequences against Helix 1}
}
\value{
a combined visual of all the helix alignments
}
\description{
A function that takes the output of createHelixAlignments() and create visual
representations of the alignments.
}
\references{
ref here
}
